Changed name of LevelEditor to MapEditor.  LevelEditor is for placing units now.
Hopefully this doesn't break things, but I refactored all the old LevelEditor stuff to be MapEditor and save into a Map folder instead.  Made resize and load level seperate buttons for MapEditor.  Added scene called LevelEditor.  It is rough right now but will be used to place units and will save into Levels.  Thought it might be useful in the future to have MapEditor and LevelEditor be subclasses to the abstract class Editor.  None of the buttons work for LevelEditor yet, but the default map loaded is DemoMap.