Added code to serialize units places in LevelEditor
When you save, the file will save with info on the map name and unit info to Levels.  Made TestLevelEditor scene to make sure deserialization works.  Seems to work.  TODO: need to allow for scrolling through new units.  Still need to add drop down to choose whether it is a player unit or enemy unit.  For some reason, delete not working.  Not registering when clicking on unit.