path manager
centralizes all references to the paths for the map and level folders into one utility class. this fixes a bug where level and map editors, or demobattle when launched directly and not from the main menu, could not load files because the streamingassets path had not yet been loaded when they were first constructed.